fix fluid tanks I think, fix actions, more fixes probably

+++ src/main/java/com/simibubi/create/content/contraptions/fluids/tank/CreativeFluidTankTileEntity.java
@@ -29,7 +29,7 @@ public class CreativeFluidTankTileEntity extends FluidTankTileEntity {
 
 	public static class CreativeSmartFluidTank extends SmartFluidTank {
 
-		public CreativeSmartFluidTank(int capacity, Consumer<FluidStack> updateCallback) {
+		public CreativeSmartFluidTank(long capacity, Consumer<FluidStack> updateCallback) {
 			super(capacity, updateCallback);
 		}
