+++ src/main/java/com/simibubi/create/foundation/fluid/SmartFluidTank.java
@@ -10,7 +10,7 @@ public class SmartFluidTank extends FluidTank {
 
 	private Consumer<FluidStack> updateCallback;
 
-	public SmartFluidTank(int capacity, Consumer<FluidStack> updateCallback) {
+	public SmartFluidTank(long capacity, Consumer<FluidStack> updateCallback) {
 		super(capacity);
 		this.updateCallback = updateCallback;
 	}
